@Shtilianova 2020-04-14
Hello @mauriciolauffer ,
Thank you for your pull request. I've created an internal record 2070154734. We will investigate your proposal and the status of the issue will be updated here in GitHub. Meanwhile please fulfill the required checks and also please add QUnit which are covering your code.
Regards, Diana