@mauriciolauffer 2020-04-15
Hi @Shtilianova 
My PR doesn't change the function signature nor what it returns, therefore there's no need for new unit tests as the change is already covered by the existing tests.

Travis-CI is failing for another reason not related to this change as you can see in its log.

![image](https://user-images.githubusercontent.com/443888/79399914-f5f64b80-7fc7-11ea-9f37-1d9d9b16bea4.png)