@Martotko 2020-04-16
Hello @mauriciolauffer ,

Please try to pull the latest changes from the master in order to pass the checks.

Best regards,
Martin