@mauriciolauffer 2020-04-17
I'm closing this one and making a new clean fork and PR on https://github.com/SAP/openui5/pull/2875